Remove boilerplate to rfl::internal::to_ptr_field_tuple
Also ensure the same behavioral changes that happened to
move/copy_to_field_tuple are actual for to_ptr_field_tuple by wrapping
in fields if necessary
