+++ include/rfl/internal/to_ptr_field_tuple.hpp
@@ -5,2326 +5,29 @@
 #include <tuple>
 #include <type_traits>
 
-#include "rfl/always_false.hpp"
-#include "rfl/internal/has_n_fields.hpp"
+#include "rfl/field_names_t.hpp"
+#include "rfl/internal/bind_to_tuple.hpp"
+#include "rfl/internal/has_fields.hpp"
 #include "rfl/internal/is_named_tuple.hpp"
 #include "rfl/internal/nt_to_ptr_named_tuple.hpp"
 #include "rfl/internal/to_ptr_field.hpp"
+#include "rfl/internal/wrap_in_fields.hpp"
 
 namespace rfl {
 namespace internal {
 
-/*The following boilerplate code was generated using a Python script:
-def make_field_template(num: int) -> str:
-    fields = ", ".join([f"f{i+1}" for i in range(num)])
-    ptr_fields = ", ".join([f"to_ptr_field(f{i+1})" for i in range(num)])
-    return (
-        """
-  } else if constexpr (has_n_fields<T, """
-        + str(num)
-        + """>) {"""
-        + ("auto& [" + fields + "] = _t;" if num > 0 else "")
-        + """
-  return std::make_tuple("""
-        + ptr_fields
-        + ");"
-    )
-
-code = "".join((make_field_template(i) for i in range(101)))
-
-with open("generated_code3.cpp", "w", encoding="utf-8") as codefile:
-    codefile.write(code)
-*/
-
 template <class T>
-auto to_ptr_field_tuple(const T& _t) {
+auto to_ptr_field_tuple(T& _t) {
   if constexpr (std::is_pointer_v<std::decay_t<T>>) {
     return to_ptr_field_tuple(*_t);
   } else if constexpr (is_named_tuple_v<T>) {
     return nt_to_ptr_named_tuple(_t).fields();
-  } else if constexpr (has_n_fields<T, 0>) {
-    return std::make_tuple();
-  } else if constexpr (has_n_fields<T, 1>) {
-    auto& [f1] = _t;
-    return std::make_tuple(to_ptr_field(f1));
-  } else if constexpr (has_n_fields<T, 2>) {
-    auto& [f1, f2] = _t;
-    return std::make_tuple(to_ptr_field(f1), to_ptr_field(f2));
-  } else if constexpr (has_n_fields<T, 3>) {
-    auto& [f1, f2, f3] = _t;
-    return std::make_tuple(to_ptr_field(f1), to_ptr_field(f2),
-                           to_ptr_field(f3));
-  } else if constexpr (has_n_fields<T, 4>) {
-    auto& [f1, f2, f3, f4] = _t;
-    return std::make_tuple(to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3),
-                           to_ptr_field(f4));
-  } else if constexpr (has_n_fields<T, 5>) {
-    auto& [f1, f2, f3, f4, f5] = _t;
-    return std::make_tuple(to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3),
-                           to_ptr_field(f4), to_ptr_field(f5));
-  } else if constexpr (has_n_fields<T, 6>) {
-    auto& [f1, f2, f3, f4, f5, f6] = _t;
-    return std::make_tuple(to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3),
-                           to_ptr_field(f4), to_ptr_field(f5),
-                           to_ptr_field(f6));
-  } else if constexpr (has_n_fields<T, 7>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7] = _t;
-    return std::make_tuple(to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3),
-                           to_ptr_field(f4), to_ptr_field(f5), to_ptr_field(f6),
-                           to_ptr_field(f7));
-  } else if constexpr (has_n_fields<T, 8>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8] = _t;
-    return std::make_tuple(to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3),
-                           to_ptr_field(f4), to_ptr_field(f5), to_ptr_field(f6),
-                           to_ptr_field(f7), to_ptr_field(f8));
-  } else if constexpr (has_n_fields<T, 9>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9] = _t;
-    return std::make_tuple(to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3),
-                           to_ptr_field(f4), to_ptr_field(f5), to_ptr_field(f6),
-                           to_ptr_field(f7), to_ptr_field(f8),
-                           to_ptr_field(f9));
-  } else if constexpr (has_n_fields<T, 10>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10] = _t;
-    return std::make_tuple(to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3),
-                           to_ptr_field(f4), to_ptr_field(f5), to_ptr_field(f6),
-                           to_ptr_field(f7), to_ptr_field(f8), to_ptr_field(f9),
-                           to_ptr_field(f10));
-  } else if constexpr (has_n_fields<T, 11>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11] = _t;
-    return std::make_tuple(to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3),
-                           to_ptr_field(f4), to_ptr_field(f5), to_ptr_field(f6),
-                           to_ptr_field(f7), to_ptr_field(f8), to_ptr_field(f9),
-                           to_ptr_field(f10), to_ptr_field(f11));
-  } else if constexpr (has_n_fields<T, 12>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12] = _t;
-    return std::make_tuple(to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3),
-                           to_ptr_field(f4), to_ptr_field(f5), to_ptr_field(f6),
-                           to_ptr_field(f7), to_ptr_field(f8), to_ptr_field(f9),
-                           to_ptr_field(f10), to_ptr_field(f11),
-                           to_ptr_field(f12));
-  } else if constexpr (has_n_fields<T, 13>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13] = _t;
-    return std::make_tuple(to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3),
-                           to_ptr_field(f4), to_ptr_field(f5), to_ptr_field(f6),
-                           to_ptr_field(f7), to_ptr_field(f8), to_ptr_field(f9),
-                           to_ptr_field(f10), to_ptr_field(f11),
-                           to_ptr_field(f12), to_ptr_field(f13));
-  } else if constexpr (has_n_fields<T, 14>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14));
-  } else if constexpr (has_n_fields<T, 15>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15] =
-        _t;
-    return std::make_tuple(to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3),
-                           to_ptr_field(f4), to_ptr_field(f5), to_ptr_field(f6),
-                           to_ptr_field(f7), to_ptr_field(f8), to_ptr_field(f9),
-                           to_ptr_field(f10), to_ptr_field(f11),
-                           to_ptr_field(f12), to_ptr_field(f13),
-                           to_ptr_field(f14), to_ptr_field(f15));
-  } else if constexpr (has_n_fields<T, 16>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16));
-  } else if constexpr (has_n_fields<T, 17>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17));
-  } else if constexpr (has_n_fields<T, 18>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18));
-  } else if constexpr (has_n_fields<T, 19>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19));
-  } else if constexpr (has_n_fields<T, 20>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20));
-  } else if constexpr (has_n_fields<T, 21>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21));
-  } else if constexpr (has_n_fields<T, 22>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22));
-  } else if constexpr (has_n_fields<T, 23>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23));
-  } else if constexpr (has_n_fields<T, 24>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24));
-  } else if constexpr (has_n_fields<T, 25>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25));
-  } else if constexpr (has_n_fields<T, 26>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26));
-  } else if constexpr (has_n_fields<T, 27>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27));
-  } else if constexpr (has_n_fields<T, 28>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28] =
-        _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28));
-  } else if constexpr (has_n_fields<T, 29>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28,
-           f29] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29));
-  } else if constexpr (has_n_fields<T, 30>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30));
-  } else if constexpr (has_n_fields<T, 31>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31));
-  } else if constexpr (has_n_fields<T, 32>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32));
-  } else if constexpr (has_n_fields<T, 33>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33));
-  } else if constexpr (has_n_fields<T, 34>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34));
-  } else if constexpr (has_n_fields<T, 35>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35));
-  } else if constexpr (has_n_fields<T, 36>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36));
-  } else if constexpr (has_n_fields<T, 37>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37));
-  } else if constexpr (has_n_fields<T, 38>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38));
-  } else if constexpr (has_n_fields<T, 39>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39));
-  } else if constexpr (has_n_fields<T, 40>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40));
-  } else if constexpr (has_n_fields<T, 41>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41));
-  } else if constexpr (has_n_fields<T, 42>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42] =
-        _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42));
-  } else if constexpr (has_n_fields<T, 43>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42,
-           f43] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43));
-  } else if constexpr (has_n_fields<T, 44>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44));
-  } else if constexpr (has_n_fields<T, 45>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45));
-  } else if constexpr (has_n_fields<T, 46>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46));
-  } else if constexpr (has_n_fields<T, 47>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47));
-  } else if constexpr (has_n_fields<T, 48>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48));
-  } else if constexpr (has_n_fields<T, 49>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49));
-  } else if constexpr (has_n_fields<T, 50>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50));
-  } else if constexpr (has_n_fields<T, 51>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51));
-  } else if constexpr (has_n_fields<T, 52>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52));
-  } else if constexpr (has_n_fields<T, 53>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53));
-  } else if constexpr (has_n_fields<T, 54>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54));
-  } else if constexpr (has_n_fields<T, 55>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55));
-  } else if constexpr (has_n_fields<T, 56>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56] =
-        _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56));
-  } else if constexpr (has_n_fields<T, 57>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56,
-           f57] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57));
-  } else if constexpr (has_n_fields<T, 58>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58));
-  } else if constexpr (has_n_fields<T, 59>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59));
-  } else if constexpr (has_n_fields<T, 60>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60));
-  } else if constexpr (has_n_fields<T, 61>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61));
-  } else if constexpr (has_n_fields<T, 62>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62));
-  } else if constexpr (has_n_fields<T, 63>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63));
-  } else if constexpr (has_n_fields<T, 64>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64));
-  } else if constexpr (has_n_fields<T, 65>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65));
-  } else if constexpr (has_n_fields<T, 66>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66));
-  } else if constexpr (has_n_fields<T, 67>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67));
-  } else if constexpr (has_n_fields<T, 68>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68));
-  } else if constexpr (has_n_fields<T, 69>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69));
-  } else if constexpr (has_n_fields<T, 70>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70] =
-        _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70));
-  } else if constexpr (has_n_fields<T, 71>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70,
-           f71] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71));
-  } else if constexpr (has_n_fields<T, 72>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72));
-  } else if constexpr (has_n_fields<T, 73>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73));
-  } else if constexpr (has_n_fields<T, 74>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73, f74] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73), to_ptr_field(f74));
-  } else if constexpr (has_n_fields<T, 75>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73, f74, f75] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73), to_ptr_field(f74),
-        to_ptr_field(f75));
-  } else if constexpr (has_n_fields<T, 76>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73, f74, f75, f76] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73), to_ptr_field(f74),
-        to_ptr_field(f75), to_ptr_field(f76));
-  } else if constexpr (has_n_fields<T, 77>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73, f74, f75, f76, f77] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73), to_ptr_field(f74),
-        to_ptr_field(f75), to_ptr_field(f76), to_ptr_field(f77));
-  } else if constexpr (has_n_fields<T, 78>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73, f74, f75, f76, f77, f78] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73), to_ptr_field(f74),
-        to_ptr_field(f75), to_ptr_field(f76), to_ptr_field(f77),
-        to_ptr_field(f78));
-  } else if constexpr (has_n_fields<T, 79>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73, f74, f75, f76, f77, f78, f79] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73), to_ptr_field(f74),
-        to_ptr_field(f75), to_ptr_field(f76), to_ptr_field(f77),
-        to_ptr_field(f78), to_ptr_field(f79));
-  } else if constexpr (has_n_fields<T, 80>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73, f74, f75, f76, f77, f78, f79, f80] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73), to_ptr_field(f74),
-        to_ptr_field(f75), to_ptr_field(f76), to_ptr_field(f77),
-        to_ptr_field(f78), to_ptr_field(f79), to_ptr_field(f80));
-  } else if constexpr (has_n_fields<T, 81>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73, f74, f75, f76, f77, f78, f79, f80, f81] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73), to_ptr_field(f74),
-        to_ptr_field(f75), to_ptr_field(f76), to_ptr_field(f77),
-        to_ptr_field(f78), to_ptr_field(f79), to_ptr_field(f80),
-        to_ptr_field(f81));
-  } else if constexpr (has_n_fields<T, 82>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73, f74, f75, f76, f77, f78, f79, f80, f81, f82] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73), to_ptr_field(f74),
-        to_ptr_field(f75), to_ptr_field(f76), to_ptr_field(f77),
-        to_ptr_field(f78), to_ptr_field(f79), to_ptr_field(f80),
-        to_ptr_field(f81), to_ptr_field(f82));
-  } else if constexpr (has_n_fields<T, 83>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73, f74, f75, f76, f77, f78, f79, f80, f81, f82, f83] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73), to_ptr_field(f74),
-        to_ptr_field(f75), to_ptr_field(f76), to_ptr_field(f77),
-        to_ptr_field(f78), to_ptr_field(f79), to_ptr_field(f80),
-        to_ptr_field(f81), to_ptr_field(f82), to_ptr_field(f83));
-  } else if constexpr (has_n_fields<T, 84>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73, f74, f75, f76, f77, f78, f79, f80, f81, f82, f83, f84] =
-        _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73), to_ptr_field(f74),
-        to_ptr_field(f75), to_ptr_field(f76), to_ptr_field(f77),
-        to_ptr_field(f78), to_ptr_field(f79), to_ptr_field(f80),
-        to_ptr_field(f81), to_ptr_field(f82), to_ptr_field(f83),
-        to_ptr_field(f84));
-  } else if constexpr (has_n_fields<T, 85>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73, f74, f75, f76, f77, f78, f79, f80, f81, f82, f83, f84,
-           f85] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73), to_ptr_field(f74),
-        to_ptr_field(f75), to_ptr_field(f76), to_ptr_field(f77),
-        to_ptr_field(f78), to_ptr_field(f79), to_ptr_field(f80),
-        to_ptr_field(f81), to_ptr_field(f82), to_ptr_field(f83),
-        to_ptr_field(f84), to_ptr_field(f85));
-  } else if constexpr (has_n_fields<T, 86>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73, f74, f75, f76, f77, f78, f79, f80, f81, f82, f83, f84, f85,
-           f86] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73), to_ptr_field(f74),
-        to_ptr_field(f75), to_ptr_field(f76), to_ptr_field(f77),
-        to_ptr_field(f78), to_ptr_field(f79), to_ptr_field(f80),
-        to_ptr_field(f81), to_ptr_field(f82), to_ptr_field(f83),
-        to_ptr_field(f84), to_ptr_field(f85), to_ptr_field(f86));
-  } else if constexpr (has_n_fields<T, 87>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73, f74, f75, f76, f77, f78, f79, f80, f81, f82, f83, f84, f85,
-           f86, f87] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73), to_ptr_field(f74),
-        to_ptr_field(f75), to_ptr_field(f76), to_ptr_field(f77),
-        to_ptr_field(f78), to_ptr_field(f79), to_ptr_field(f80),
-        to_ptr_field(f81), to_ptr_field(f82), to_ptr_field(f83),
-        to_ptr_field(f84), to_ptr_field(f85), to_ptr_field(f86),
-        to_ptr_field(f87));
-  } else if constexpr (has_n_fields<T, 88>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73, f74, f75, f76, f77, f78, f79, f80, f81, f82, f83, f84, f85,
-           f86, f87, f88] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73), to_ptr_field(f74),
-        to_ptr_field(f75), to_ptr_field(f76), to_ptr_field(f77),
-        to_ptr_field(f78), to_ptr_field(f79), to_ptr_field(f80),
-        to_ptr_field(f81), to_ptr_field(f82), to_ptr_field(f83),
-        to_ptr_field(f84), to_ptr_field(f85), to_ptr_field(f86),
-        to_ptr_field(f87), to_ptr_field(f88));
-  } else if constexpr (has_n_fields<T, 89>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73, f74, f75, f76, f77, f78, f79, f80, f81, f82, f83, f84, f85,
-           f86, f87, f88, f89] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73), to_ptr_field(f74),
-        to_ptr_field(f75), to_ptr_field(f76), to_ptr_field(f77),
-        to_ptr_field(f78), to_ptr_field(f79), to_ptr_field(f80),
-        to_ptr_field(f81), to_ptr_field(f82), to_ptr_field(f83),
-        to_ptr_field(f84), to_ptr_field(f85), to_ptr_field(f86),
-        to_ptr_field(f87), to_ptr_field(f88), to_ptr_field(f89));
-  } else if constexpr (has_n_fields<T, 90>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73, f74, f75, f76, f77, f78, f79, f80, f81, f82, f83, f84, f85,
-           f86, f87, f88, f89, f90] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73), to_ptr_field(f74),
-        to_ptr_field(f75), to_ptr_field(f76), to_ptr_field(f77),
-        to_ptr_field(f78), to_ptr_field(f79), to_ptr_field(f80),
-        to_ptr_field(f81), to_ptr_field(f82), to_ptr_field(f83),
-        to_ptr_field(f84), to_ptr_field(f85), to_ptr_field(f86),
-        to_ptr_field(f87), to_ptr_field(f88), to_ptr_field(f89),
-        to_ptr_field(f90));
-  } else if constexpr (has_n_fields<T, 91>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73, f74, f75, f76, f77, f78, f79, f80, f81, f82, f83, f84, f85,
-           f86, f87, f88, f89, f90, f91] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73), to_ptr_field(f74),
-        to_ptr_field(f75), to_ptr_field(f76), to_ptr_field(f77),
-        to_ptr_field(f78), to_ptr_field(f79), to_ptr_field(f80),
-        to_ptr_field(f81), to_ptr_field(f82), to_ptr_field(f83),
-        to_ptr_field(f84), to_ptr_field(f85), to_ptr_field(f86),
-        to_ptr_field(f87), to_ptr_field(f88), to_ptr_field(f89),
-        to_ptr_field(f90), to_ptr_field(f91));
-  } else if constexpr (has_n_fields<T, 92>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73, f74, f75, f76, f77, f78, f79, f80, f81, f82, f83, f84, f85,
-           f86, f87, f88, f89, f90, f91, f92] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73), to_ptr_field(f74),
-        to_ptr_field(f75), to_ptr_field(f76), to_ptr_field(f77),
-        to_ptr_field(f78), to_ptr_field(f79), to_ptr_field(f80),
-        to_ptr_field(f81), to_ptr_field(f82), to_ptr_field(f83),
-        to_ptr_field(f84), to_ptr_field(f85), to_ptr_field(f86),
-        to_ptr_field(f87), to_ptr_field(f88), to_ptr_field(f89),
-        to_ptr_field(f90), to_ptr_field(f91), to_ptr_field(f92));
-  } else if constexpr (has_n_fields<T, 93>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73, f74, f75, f76, f77, f78, f79, f80, f81, f82, f83, f84, f85,
-           f86, f87, f88, f89, f90, f91, f92, f93] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73), to_ptr_field(f74),
-        to_ptr_field(f75), to_ptr_field(f76), to_ptr_field(f77),
-        to_ptr_field(f78), to_ptr_field(f79), to_ptr_field(f80),
-        to_ptr_field(f81), to_ptr_field(f82), to_ptr_field(f83),
-        to_ptr_field(f84), to_ptr_field(f85), to_ptr_field(f86),
-        to_ptr_field(f87), to_ptr_field(f88), to_ptr_field(f89),
-        to_ptr_field(f90), to_ptr_field(f91), to_ptr_field(f92),
-        to_ptr_field(f93));
-  } else if constexpr (has_n_fields<T, 94>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73, f74, f75, f76, f77, f78, f79, f80, f81, f82, f83, f84, f85,
-           f86, f87, f88, f89, f90, f91, f92, f93, f94] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73), to_ptr_field(f74),
-        to_ptr_field(f75), to_ptr_field(f76), to_ptr_field(f77),
-        to_ptr_field(f78), to_ptr_field(f79), to_ptr_field(f80),
-        to_ptr_field(f81), to_ptr_field(f82), to_ptr_field(f83),
-        to_ptr_field(f84), to_ptr_field(f85), to_ptr_field(f86),
-        to_ptr_field(f87), to_ptr_field(f88), to_ptr_field(f89),
-        to_ptr_field(f90), to_ptr_field(f91), to_ptr_field(f92),
-        to_ptr_field(f93), to_ptr_field(f94));
-  } else if constexpr (has_n_fields<T, 95>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73, f74, f75, f76, f77, f78, f79, f80, f81, f82, f83, f84, f85,
-           f86, f87, f88, f89, f90, f91, f92, f93, f94, f95] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73), to_ptr_field(f74),
-        to_ptr_field(f75), to_ptr_field(f76), to_ptr_field(f77),
-        to_ptr_field(f78), to_ptr_field(f79), to_ptr_field(f80),
-        to_ptr_field(f81), to_ptr_field(f82), to_ptr_field(f83),
-        to_ptr_field(f84), to_ptr_field(f85), to_ptr_field(f86),
-        to_ptr_field(f87), to_ptr_field(f88), to_ptr_field(f89),
-        to_ptr_field(f90), to_ptr_field(f91), to_ptr_field(f92),
-        to_ptr_field(f93), to_ptr_field(f94), to_ptr_field(f95));
-  } else if constexpr (has_n_fields<T, 96>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73, f74, f75, f76, f77, f78, f79, f80, f81, f82, f83, f84, f85,
-           f86, f87, f88, f89, f90, f91, f92, f93, f94, f95, f96] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73), to_ptr_field(f74),
-        to_ptr_field(f75), to_ptr_field(f76), to_ptr_field(f77),
-        to_ptr_field(f78), to_ptr_field(f79), to_ptr_field(f80),
-        to_ptr_field(f81), to_ptr_field(f82), to_ptr_field(f83),
-        to_ptr_field(f84), to_ptr_field(f85), to_ptr_field(f86),
-        to_ptr_field(f87), to_ptr_field(f88), to_ptr_field(f89),
-        to_ptr_field(f90), to_ptr_field(f91), to_ptr_field(f92),
-        to_ptr_field(f93), to_ptr_field(f94), to_ptr_field(f95),
-        to_ptr_field(f96));
-  } else if constexpr (has_n_fields<T, 97>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73, f74, f75, f76, f77, f78, f79, f80, f81, f82, f83, f84, f85,
-           f86, f87, f88, f89, f90, f91, f92, f93, f94, f95, f96, f97] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73), to_ptr_field(f74),
-        to_ptr_field(f75), to_ptr_field(f76), to_ptr_field(f77),
-        to_ptr_field(f78), to_ptr_field(f79), to_ptr_field(f80),
-        to_ptr_field(f81), to_ptr_field(f82), to_ptr_field(f83),
-        to_ptr_field(f84), to_ptr_field(f85), to_ptr_field(f86),
-        to_ptr_field(f87), to_ptr_field(f88), to_ptr_field(f89),
-        to_ptr_field(f90), to_ptr_field(f91), to_ptr_field(f92),
-        to_ptr_field(f93), to_ptr_field(f94), to_ptr_field(f95),
-        to_ptr_field(f96), to_ptr_field(f97));
-  } else if constexpr (has_n_fields<T, 98>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73, f74, f75, f76, f77, f78, f79, f80, f81, f82, f83, f84, f85,
-           f86, f87, f88, f89, f90, f91, f92, f93, f94, f95, f96, f97, f98] =
-        _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73), to_ptr_field(f74),
-        to_ptr_field(f75), to_ptr_field(f76), to_ptr_field(f77),
-        to_ptr_field(f78), to_ptr_field(f79), to_ptr_field(f80),
-        to_ptr_field(f81), to_ptr_field(f82), to_ptr_field(f83),
-        to_ptr_field(f84), to_ptr_field(f85), to_ptr_field(f86),
-        to_ptr_field(f87), to_ptr_field(f88), to_ptr_field(f89),
-        to_ptr_field(f90), to_ptr_field(f91), to_ptr_field(f92),
-        to_ptr_field(f93), to_ptr_field(f94), to_ptr_field(f95),
-        to_ptr_field(f96), to_ptr_field(f97), to_ptr_field(f98));
-  } else if constexpr (has_n_fields<T, 99>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73, f74, f75, f76, f77, f78, f79, f80, f81, f82, f83, f84, f85,
-           f86, f87, f88, f89, f90, f91, f92, f93, f94, f95, f96, f97, f98,
-           f99] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73), to_ptr_field(f74),
-        to_ptr_field(f75), to_ptr_field(f76), to_ptr_field(f77),
-        to_ptr_field(f78), to_ptr_field(f79), to_ptr_field(f80),
-        to_ptr_field(f81), to_ptr_field(f82), to_ptr_field(f83),
-        to_ptr_field(f84), to_ptr_field(f85), to_ptr_field(f86),
-        to_ptr_field(f87), to_ptr_field(f88), to_ptr_field(f89),
-        to_ptr_field(f90), to_ptr_field(f91), to_ptr_field(f92),
-        to_ptr_field(f93), to_ptr_field(f94), to_ptr_field(f95),
-        to_ptr_field(f96), to_ptr_field(f97), to_ptr_field(f98),
-        to_ptr_field(f99));
-  } else if constexpr (has_n_fields<T, 100>) {
-    auto& [f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15,
-           f16, f17, f18, f19, f20, f21, f22, f23, f24, f25, f26, f27, f28, f29,
-           f30, f31, f32, f33, f34, f35, f36, f37, f38, f39, f40, f41, f42, f43,
-           f44, f45, f46, f47, f48, f49, f50, f51, f52, f53, f54, f55, f56, f57,
-           f58, f59, f60, f61, f62, f63, f64, f65, f66, f67, f68, f69, f70, f71,
-           f72, f73, f74, f75, f76, f77, f78, f79, f80, f81, f82, f83, f84, f85,
-           f86, f87, f88, f89, f90, f91, f92, f93, f94, f95, f96, f97, f98, f99,
-           f100] = _t;
-    return std::make_tuple(
-        to_ptr_field(f1), to_ptr_field(f2), to_ptr_field(f3), to_ptr_field(f4),
-        to_ptr_field(f5), to_ptr_field(f6), to_ptr_field(f7), to_ptr_field(f8),
-        to_ptr_field(f9), to_ptr_field(f10), to_ptr_field(f11),
-        to_ptr_field(f12), to_ptr_field(f13), to_ptr_field(f14),
-        to_ptr_field(f15), to_ptr_field(f16), to_ptr_field(f17),
-        to_ptr_field(f18), to_ptr_field(f19), to_ptr_field(f20),
-        to_ptr_field(f21), to_ptr_field(f22), to_ptr_field(f23),
-        to_ptr_field(f24), to_ptr_field(f25), to_ptr_field(f26),
-        to_ptr_field(f27), to_ptr_field(f28), to_ptr_field(f29),
-        to_ptr_field(f30), to_ptr_field(f31), to_ptr_field(f32),
-        to_ptr_field(f33), to_ptr_field(f34), to_ptr_field(f35),
-        to_ptr_field(f36), to_ptr_field(f37), to_ptr_field(f38),
-        to_ptr_field(f39), to_ptr_field(f40), to_ptr_field(f41),
-        to_ptr_field(f42), to_ptr_field(f43), to_ptr_field(f44),
-        to_ptr_field(f45), to_ptr_field(f46), to_ptr_field(f47),
-        to_ptr_field(f48), to_ptr_field(f49), to_ptr_field(f50),
-        to_ptr_field(f51), to_ptr_field(f52), to_ptr_field(f53),
-        to_ptr_field(f54), to_ptr_field(f55), to_ptr_field(f56),
-        to_ptr_field(f57), to_ptr_field(f58), to_ptr_field(f59),
-        to_ptr_field(f60), to_ptr_field(f61), to_ptr_field(f62),
-        to_ptr_field(f63), to_ptr_field(f64), to_ptr_field(f65),
-        to_ptr_field(f66), to_ptr_field(f67), to_ptr_field(f68),
-        to_ptr_field(f69), to_ptr_field(f70), to_ptr_field(f71),
-        to_ptr_field(f72), to_ptr_field(f73), to_ptr_field(f74),
-        to_ptr_field(f75), to_ptr_field(f76), to_ptr_field(f77),
-        to_ptr_field(f78), to_ptr_field(f79), to_ptr_field(f80),
-        to_ptr_field(f81), to_ptr_field(f82), to_ptr_field(f83),
-        to_ptr_field(f84), to_ptr_field(f85), to_ptr_field(f86),
-        to_ptr_field(f87), to_ptr_field(f88), to_ptr_field(f89),
-        to_ptr_field(f90), to_ptr_field(f91), to_ptr_field(f92),
-        to_ptr_field(f93), to_ptr_field(f94), to_ptr_field(f95),
-        to_ptr_field(f96), to_ptr_field(f97), to_ptr_field(f98),
-        to_ptr_field(f99), to_ptr_field(f100));
-
-    // ---------------------------------
-    // End of generated boilerplate code
-    // ---------------------------------
+  } else if constexpr (has_fields<T>()) {
+    return bind_to_tuple(_t, [](auto& x) { return to_ptr_field(x); });
   } else {
-    static_assert(
-        rfl::always_false_v<T>,
-        "\n\nThis error occurs for one of two reasons:\n\n"
-        "1) You have created a struct with more than 100 fields, which is "
-        "unsupported. Please split up your struct into several "
-        "smaller structs and then use rfl::Flatten<...> to combine them. "
-        "Refer "
-        "to the documentation on rfl::Flatten<...> for details.\n\n"
-        "2) You have added a custom constructor to your struct, which you "
-        "shouldn't do either. Please refer to the sections on custom "
-        "classes or custom parsers in the documentation "
-        "for solutions to this problem.\n\n");
+    using FieldNames = field_names_t<T>;
+    auto tup = bind_to_tuple(_t, [](auto& x) { return to_ptr_field(x); });
+    return wrap_in_fields<FieldNames>(std::move(tup));
   }
 }
